requirements.txt
opencv-python
pillow
pytesseract
pandas
numpy
pdf2image
xlsxwriter
openpyxl
pyinstaller